third_party/blink/web_tests/fast/xpath/nsresolver-exception-expected.txt

CONSOLE ERROR: Uncaught Error: this is an exception
SUCCESS
